Update disk space when Flight/Project is hard deleted
