+++ core/models.py
@@ -454,6 +454,7 @@ def delete_geoserver_workspace(sender, instance: Union[Flight, UserProject], **k
 
 def delete_on_disk(sender, instance: Union[Flight, UserProject], **kwargs):
     shutil.rmtree(instance.get_disk_path())
+    instance.user.update_disk_space()
 
 
 def delete_thumbnail(sender, instance: Flight, **kwargs):
